Feature/fog resolver refactor (#2452)
* wip refactoring into separate crates

* continued reorg and cleanup of code

* adding docs

* updating cargo lock

* removing default from cargo

* updating use of FogResolver

* updating FogReportTypes usage

* incorrect cargo crate fix

* removing dupe import of FogResolver

* fixing incorrect crate import

* remove unused import

* Update fog/report/resolver/Cargo.toml


* updating comment
