+++ fog/distribution/Cargo.toml
@@ -19,7 +19,7 @@ mc-consensus-enclave-measurement = { path = "../../consensus/enclave/measurement
 mc-crypto-keys = { path = "../../crypto/keys" }
 mc-fog-ingest-enclave-measurement = { path = "../ingest/enclave/measurement" }
 mc-fog-report-connection = { path = "../../fog/report/connection" }
-mc-fog-report-validation = { path = "../../fog/report/validation" }
+mc-fog-report-resolver = { path = "../../fog/report/resolver" }
 mc-ledger-db = { path = "../../ledger/db" }
 mc-transaction-core = { path = "../../transaction/core" }
 mc-transaction-std = { path = "../../transaction/std" }
