+++ fog/report/validation/src/lib.rs
@@ -14,127 +14,104 @@
 
 extern crate alloc;
 
-mod traits;
-
-/// Data structure for fog-ingest report validation
-pub mod ingest_report;
-
-#[cfg(feature = "automock")]
-pub use crate::traits::MockFogPubkeyResolver;
-pub use crate::traits::{FogPubkeyError, FogPubkeyResolver, FullyValidatedFogPubkey};
-
-use crate::ingest_report::IngestReportVerifier;
-use alloc::{
-    collections::BTreeMap,
-    string::{String, ToString},
-};
-use core::str::FromStr;
+use core::fmt::{Debug, Display};
+use displaydoc::Display;
 use mc_account_keys::PublicAddress;
-use mc_attest_verifier::Verifier;
-use mc_fog_report_types::ReportResponse;
-use mc_fog_sig::Verifier as FogSigVerifier;
-use mc_util_uri::{FogUri, UriParseError};
-use serde::{Deserialize, Serialize};
+use mc_crypto_keys::RistrettoPublic;
+use mc_fog_sig::Error as FogSigError;
+use mc_util_uri::UriParseError;
+#[cfg(feature = "automock")]
+use mockall::*;
+use std::collections::HashMap;
 
-/// Represents a set of unvalidated responses from Fog report servers
-/// Key = Fog-url that was contacted, must match the string in user's public
-/// address Value = The complete response from the fog report server
-///
-/// When constructing a transaction, the fog-url for each recipient should be
-/// extracted from their public address, then a request to that report server
-/// should be made. The responses should be collected in a map-structure (like
-/// this). This should be done for each recipient.
-///
-/// This map structure is ultimately consumed by the TransactionBuilder object,
-/// which validates the responses against the fog data in the public addresses
-/// when building the transaction.
-///
-/// This map structure should not be cached, because the fog pubkeys have an
-/// expiry date and don't live that long. They can be cached for a short time,
-/// but the transaction builder enforces that the tombstone block for the
-/// transaction is limited by the pubkey expiry value of any fog pubkey that is
-/// used, so if these are cached too long, the transaction will be rejected by
-/// consensus.
-///
-/// In the case of constructing off-line transactions with Fog recipients, the
-/// flow is: (1) Take fog-urls from (offline) public addresses to the online
-/// machine (2) Hit the fog report servers (online machine), producing
-/// FogReportResponses (3) Take FogReportResponses to the offline machine, and
-/// use with transaction builder,     to create the transaction offline.
-/// (4) Take the constructed transaction to the online machine and submit to
-/// consensus.
-///
-/// Note: there is no particular reason for this to be BTreeMap instead of
-/// HashMap, except that it is slightly more portable, only requiring the alloc
-/// crate.
-pub type FogReportResponses = BTreeMap<String, ReportResponse>;
-
-/// A collection of unvalidated fog reports, together with an IAS verifier.
-/// This object is passed to the TransactionBuilder object.
-/// When fog is not involved, it can simply be defaulted.
-///
-/// Once constructed, this object can get validated fog pubkeys to build fog
-/// hints for transactions, without talking to the internet, and so is
-/// compatible with offline transactions to fog recipients. Only getting the
-/// FogReportResponses requires an online connection.
-#[derive(Default, Clone, Debug, Serialize, Deserialize)]
-pub struct FogResolver {
-    responses: FogReportResponses,
-    verifier: IngestReportVerifier,
+/// Class that can resolve a public address to a fully-validated fog public key
+/// structure, including the pubkey expiry data from the report server.
+#[cfg_attr(feature = "automock", automock)]
+pub trait FogPubkeyResolver {
+    /// Fetch and validate a fog public key, given a recipient's public address
+    fn get_fog_pubkey(
+        &self,
+        recipient: &PublicAddress,
+    ) -> Result<FullyValidatedFogPubkey, FogPubkeyError>;
 }
 
+/// Represents a fog public key validated to use for creating encrypted fog
+/// hints. This object should be constructed only when the IAS report has been
+/// validated, and the chain of trust from the connection has been validated,
+/// and the the fog user's fog_authority_sig over the root subjectPublicKeyInfo
+/// in the signature chain has been validated.
+#[derive(Clone, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
+pub struct FullyValidatedFogPubkey {
+    /// The ristretto curve point which was extracted from the IAS report
+    /// additional data after validation. This is the encryption key used to
+    /// create encrypted fog hints. The corresponding private key lives only
+    /// in SGX ingest nodes.
+    pub pubkey: RistrettoPublic,
+    /// The pubkey_expiry value is the latest block that fog-service promises
+    /// that is valid to encrypt fog hints using this key for.
+    /// The client should obey this limit by not setting tombstone block for a
+    /// transaction larger than this limit if the fog pubkey is used.
+    pub pubkey_expiry: u64,
+}
+
+/// An error that can occur when trying to get a validated fog pubkey from the
+/// FogResolver object
+/// TODO: Add x509 errors, user-sig check errors, etc. here
+#[derive(Display, Debug)]
+pub enum FogPubkeyError {
+    /// No matching reports response for url = {0}
+    NoMatchingReportResponse(String),
+    /// No matching report id for url = {0}, report_id = {1}
+    NoMatchingReportId(String, String),
+    /// Address has no fog_report_url, cannot fetch fog pubkey
+    NoFogReportUrl,
+    /// Failed to parse fog url: {0}
+    Url(UriParseError),
+    /// Ingest report deserialization error: {0},
+    Deserialization(mc_util_serial::decode::Error),
+    /// Ingest report verification error: {0}
+    IngestReport(String),
+    /// Authority verification error: {0}
+    Authority(String),
+}
+
+impl From<mc_util_serial::decode::Error> for FogPubkeyError {
+    fn from(src: mc_util_serial::decode::Error) -> Self {
+        Self::Deserialization(src)
+    }
+}
+
+impl From<UriParseError> for FogPubkeyError {
+    fn from(src: UriParseError) -> Self {
+        Self::Url(src)
+    }
+}
+
+impl<A: Debug + Display, R: Debug + Display> From<FogSigError<A, R>> for FogPubkeyError {
+    fn from(src: FogSigError<A, R>) -> Self {
+        FogPubkeyError::Authority(src.to_string())
+    }
+}
+
+/// A basic implementation of the FogPubkeyResolver trait that must be seeded
+/// with a HashMap of PublicAddresses to FullValidatedFogPubkeys.
+pub struct FogResolver(HashMap<PublicAddress, FullyValidatedFogPubkey>);
+
 impl FogResolver {
-    /// Create a new FogResolver object, given serialized (unverified)
-    /// fog report server responses,
-    /// and an attestation verifier for fog ingest measurements.
-    pub fn new(responses: FogReportResponses, verifier: &Verifier) -> Result<Self, UriParseError> {
-        // Normalize URI strings
-        let responses: FogReportResponses = responses
-            .into_iter()
-            .map(
-                |(uri_str, resp)| -> Result<(String, ReportResponse), UriParseError> {
-                    let uri = FogUri::from_str(&uri_str)?.to_string();
-                    Ok((uri, resp))
-                },
-            )
-            .collect::<Result<_, UriParseError>>()?;
-        Ok(Self {
-            responses,
-            verifier: IngestReportVerifier::from(verifier),
-        })
+    /// Create a new FogResolver with an filled HashMap of responses
+    pub fn new(responses: HashMap<PublicAddress, FullyValidatedFogPubkey>) -> Self {
+        FogResolver(responses)
     }
 }
 
 impl FogPubkeyResolver for FogResolver {
     fn get_fog_pubkey(
         &self,
-        recipient: &PublicAddress,
+        address: &PublicAddress,
     ) -> Result<FullyValidatedFogPubkey, FogPubkeyError> {
-        if let Some(url) = recipient.fog_report_url() {
-            // Normalize the string to URL before lookup
-            let url = FogUri::from_str(url)?.to_string();
-            if let Some(result) = self.responses.get(&url) {
-                // Verify the authority signature chain
-                recipient.verify_fog_sig(result)?;
-                // Get the report corresponding to our ID
-                let report_id = recipient.fog_report_id().unwrap_or("").to_string();
-                for report in result.reports.iter() {
-                    if report_id == report.fog_report_id {
-                        let pubkey = self
-                            .verifier
-                            .validate_ingest_ias_report(report.report.clone())?;
-                        return Ok(FullyValidatedFogPubkey {
-                            pubkey,
-                            pubkey_expiry: report.pubkey_expiry,
-                        });
-                    }
-                }
-                Err(FogPubkeyError::NoMatchingReportId(url, report_id))
-            } else {
-                Err(FogPubkeyError::NoMatchingReportResponse(url))
-            }
-        } else {
-            Err(FogPubkeyError::NoFogReportUrl)
-        }
+        self.0
+            .get(address)
+            .cloned()
+            .ok_or(FogPubkeyError::NoFogReportUrl)
     }
 }
