+++ libmobilecoin/src/error.rs
@@ -9,7 +9,7 @@ use mc_crypto_box::{AeadError, Error as CryptoBoxError};
 use mc_crypto_keys::KeyError;
 use mc_crypto_noise::CipherError;
 use mc_fog_kex_rng::Error as FogKexRngError;
-use mc_fog_report_validation::{ingest_report::Error as IngestReportError, FogPubkeyError};
+use mc_fog_report_validation::FogPubkeyError;
 use mc_transaction_core::{AmountError, BlockVersionError};
 use mc_transaction_std::TxBuilderError;
 use mc_util_serial::DecodeError;
@@ -195,14 +195,15 @@ impl From<ProtobufError> for LibMcError {
 
 impl From<TxBuilderError> for LibMcError {
     fn from(err: TxBuilderError) -> Self {
-        if let TxBuilderError::FogPublicKey(FogPubkeyError::IngestReport(
-            IngestReportError::Verifier(VerifierError::Verification(_)),
-        )) = err
-        {
-            LibMcError::AttestationVerificationFailed(format!("{:?}", err))
-        } else {
+        // TODO
+        // if let TxBuilderError::FogPublicKey(FogPubkeyError::IngestReport(
+        //     IngestReportError::Verifier(VerifierError::Verification(_)),
+        // )) = err
+        // {
+        //     LibMcError::AttestationVerificationFailed(format!("{:?}", err))
+        // } else {
             LibMcError::InvalidInput(format!("{:?}", err))
-        }
+        // }
     }
 }
 
@@ -213,9 +214,10 @@ impl From<FogPubkeyError> for LibMcError {
             | FogPubkeyError::Url(_)
             | FogPubkeyError::Deserialization(_) => LibMcError::InvalidInput(err.to_string()),
 
-            FogPubkeyError::IngestReport(IngestReportError::Verifier(
-                VerifierError::Verification(_),
-            )) => LibMcError::AttestationVerificationFailed(err.to_string()),
+            // TODO
+            // FogPubkeyError::IngestReport(IngestReportError::Verifier(
+            //     VerifierError::Verification(_),
+            // )) => LibMcError::AttestationVerificationFailed(err.to_string()),
 
             _ => LibMcError::FogPubkey(err.to_string()),
         }
