+++ libmobilecoin/src/fog.rs
@@ -7,9 +7,11 @@ use mc_account_keys::PublicAddress;
 use mc_attest_verifier::Verifier;
 use mc_crypto_keys::{ReprBytes, RistrettoPrivate, RistrettoPublic};
 use mc_fog_kex_rng::{BufferedRng, KexRngPubkey, NewFromKex, StoredRng, VersionedKexRng};
+use mc_fog_report_types::{FogReportResponses};
 use mc_fog_report_validation::{
-    FogPubkeyResolver, FogReportResponses, FogResolver, FullyValidatedFogPubkey,
+    FogPubkeyResolver, FullyValidatedFogPubkey,
 };
+use mc_fog_report_resolver::FogResolver;
 use mc_util_ffi::*;
 use mc_util_serial::Message;
 use mc_util_uri::FogUri;
