+++ mobilecoind/Cargo.toml
@@ -28,6 +28,7 @@ mc-crypto-hashes = { path = "../crypto/hashes" }
 mc-crypto-keys = { path = "../crypto/keys" }
 mc-crypto-rand = { path = "../crypto/rand" }
 mc-fog-report-connection = { path = "../fog/report/connection" }
+mc-fog-report-resolver = { path = "../fog/report/resolver" }
 mc-fog-report-validation = { path = "../fog/report/validation" }
 mc-ledger-db = { path = "../ledger/db" }
 mc-ledger-migration = { path = "../ledger/migration" }
